refactor(ui): move header/footer components to _private convention
