+++ apps/frontend/components.d.ts
@@ -14,5 +14,6 @@ declare module 'vue' {
     Button: typeof import('primevue/button')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
+    Skeleton: typeof import('primevue/skeleton')['default']
   }
 }
