+++ apps/frontend/src/common/ui/DefaultLayout.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
-import Footer from "./footer/Footer.vue";
-import Header from "./header/Header.vue";
+import Footer from "./Footer.vue";
+import Header from "./Header.vue";
 </script>
 
 <template>
   <Header />
 
-  <slot></slot>
+  <slot />
 
   <Footer />
 </template>
