+++ apps/frontend/src/common/ui/_footer/FooterCopyright.vue
