+++ apps/frontend/src/common/ui/_footer/FooterSocials.vue
